requirements.txt
flask==3.0.3
slack-sdk==3.31.0
python-dotenv==1.0.1
gunicorn==22.0.0
